update amend and search cert
