+++ cypress/support/searchNoticeRegistry.cy.js
@@ -1,5 +1,5 @@
 import { lraTransaction } from "./lraTransactionProcess.cy";
-import { searchCertReg  } from "./searchCertRegistry.cy";
+import { searchCertReg } from "./searchCertRegistry.cy";
 
 
 export function searchNoticeReg() {
@@ -20,29 +20,37 @@ export function searchNoticeReg() {
 
     cy.xpath(testdata.search).click();
     lraTransaction();
-   // cy.xpath(testdata.okay).click();
+    // cy.xpath(testdata.okay).click();
 
-     //search by Notice Registration
-     cy.contains('Search Notice').click();
-     cy.contains('Proceed').click();
-     cy.xpath(testdata.e).click();
- 
-     cy.xpath(testdata.f).type('N230100019915');
-     cy.xpath(testdata.searchReg).click();
-     cy.wait(5000);
-    lraTransaction();
-  //  cy.xpath(testdata.okay).click();
+    //--
 
-     //search by collateral details, Motor Vehicle
-     cy.contains('Search Notice').click();
-     cy.contains('Proceed').click();
-     cy.xpath(testdata.g).click();
-     cy.xpath(testdata.h).type('chassis1');
-     cy.xpath(testdata.searchCol).click();
-     cy.wait(5000);
+
+    //--
+    //  cy.xpath(testdata.okay).click();
+
+    //search by collateral details, Motor Vehicle
+    cy.contains('Search Notice').click();
+    cy.contains('Proceed').click();
+    cy.xpath(testdata.g).click();
+    cy.xpath(testdata.h).type('chassis1');
+    cy.xpath(testdata.searchCol).click();
+    cy.wait(5000);
     lraTransaction();
-   // cy.xpath(testdata.okay).click();
-   
+    // cy.xpath(testdata.okay).click();
+
+    
+    //search by Notice Registration
+    //  cy.contains('Search the Registry').click();
+    cy.contains('Proceed').click();
+    cy.contains('Search Notice').click();
+    //  cy.contains('Proceed').click();
+    cy.xpath(testdata.e).click();
+
+    cy.xpath(testdata.f).type('N221100000019');
+    cy.xpath(testdata.searchReg).click({ timeout: 10000 });
+
+    lraTransaction();
+
     searchCertReg();
 }
 
@@ -66,7 +74,7 @@ class testData {
     search = '//*[@id="__layout"]/section/main/div/div/div[1]/div[2]/div/form/div/div[7]/div/div/button/span';
 
     noticeReg = '//*[@id="__layout"]/section/main/div/div/div[1]/div[1]/div[2]/label/span[1]';
-    
+
     searchReg = '//*[@id="__layout"]/section/main/div/div/div[1]/div[2]/div/form/div/div[2]/div/div/button/span';
     searchCol = '//*[@id="__layout"]/section/main/div/div/div[1]/div[2]/div/div/div[2]/form/div[2]/div/button/span';
 }
